Moving concat with props logic to merge-styles.
