+++ packages/utilities/src/styled.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { concatStyleSets, IStyleSet, IStyleFunctionOrObject } from '@uifabric/merge-styles';
+import { IStyleSet, IStyleFunctionOrObject, concatStyleSetsWithProps } from '@uifabric/merge-styles';
 import { Customizations } from './customizations/Customizations';
 import { CustomizerContext, ICustomizerContext } from './customizations/CustomizerContext';
 
@@ -101,7 +101,7 @@ export function styled<
         // this._customizedStyles = customizedStyles;
 
         // Using styled components as the Component arg will result in nested styling arrays.
-        this._styles = (styleProps: TStyleProps) => _resolve(styleProps, baseStyles, customizedStyles, this.props.styles);
+        this._styles = (styleProps: TStyleProps) => concatStyleSetsWithProps(styleProps, baseStyles, customizedStyles, this.props.styles);
 
         // The __cachedInputs__ array is attached to the function and consumed by the
         // classNamesFunction as a list of keys to include for memoizing classnames.
@@ -117,30 +117,4 @@ export function styled<
   // This preserves backwards compatibility.
   // tslint:disable-next-line:no-any
   return Wrapped as any;
-}
-
-function _resolve<TStyleProps, TStyleSet extends IStyleSet<TStyleSet>>(
-  styleProps: TStyleProps,
-  ...allStyles: (IStyleFunctionOrObject<TStyleProps, TStyleSet> | undefined)[]
-): Partial<TStyleSet> {
-  const result: Partial<TStyleSet>[] = [];
-
-  for (const styles of allStyles) {
-    if (styles) {
-      result.push(typeof styles === 'function' ? styles(styleProps) : styles);
-    }
-  }
-
-  if (result.length === 1) {
-    return result[0] as Partial<TStyleSet>;
-  } else if (result.length) {
-    // cliffkoh: I cannot figure out how to avoid the cast to any here.
-    // It is something to do with the use of Omit in IStyleSet.
-    // It might not be necessary once  Omit becomes part of lib.d.ts (when we remove our own Omit and rely on
-    // the official version).
-    // tslint:disable-next-line:no-any
-    return concatStyleSets(...(result as any)) as any;
-  }
-
-  return {};
 }
